initial commit
Fyyur is a musical venue and artist booking site that facilitates the discovery and bookings of shows between local performing artists and venues. I built this as a submission entry for a Udacity Full-stack course I took. This site lets you list new artists and venues, discover them, and list shows with artists as a venue owner.